+++ tests/error/result_errors/state_result_tag_too_long.py
@@ -1,5 +1,5 @@
 from guppylang.std.builtins import comptime
-from guppylang.std.debug import state_result
+from guppylang.std.debug import state_output
 from guppylang.std.quantum import discard, qubit
 from tests.util import compile_guppy
 
@@ -8,5 +8,5 @@ TAG_MAX_LEN = 200
 @compile_guppy
 def main() -> None:
     q1 = qubit()
-    state_result(comptime("a" * (TAG_MAX_LEN + 1)), q1)
+    state_output(comptime("a" * (TAG_MAX_LEN + 1)), q1)
     discard(q1)