+++ tests/integration/library/test_entrypoint.py
@@ -1,9 +1,11 @@
 import pytest
 
+from hugr._hugr.linking import HugrLinkingError
+
 from guppylang import guppy
 from guppylang.defs import GuppyDefinition
 from guppylang.emulator import EmulatorBuilder
-from guppylang.std.platform import result
+from guppylang.std.platform import output
 
 
 def test_manual_link_no_entrypoints():
@@ -34,7 +36,7 @@ def test_manual_link_entrypoint_lhs():
 
     @guppy
     def main() -> None:
-        result("result", decl(5))
+        output("result", decl(5))
 
     linked = main.compile().link(adder_lib)
     emulator = EmulatorBuilder().build(linked, n_qubits=1)
@@ -53,7 +55,7 @@ def test_manual_link_entrypoint_rhs():
 
     @guppy
     def main() -> None:
-        result("result", decl(5))
+        output("result", decl(5))
 
     linked = adder_lib.link(main.compile())
     emulator = EmulatorBuilder().build(linked, n_qubits=1)
@@ -64,12 +66,15 @@ def test_manual_link_multiple_entrypoints():
     def produce_entrypoint() -> GuppyDefinition:
         @guppy
         def main() -> None:
-            result("result", 1)
+            output("result", 1)
 
         return main
 
     lib1 = produce_entrypoint().compile()
     lib2 = produce_entrypoint().compile()
 
-    with pytest.raises(ValueError, match="Cannot link two executable modules together"):
+    with pytest.raises(
+        HugrLinkingError,
+        match="Cannot link two modules with non-root entrypoints together",
+    ):
         lib1.link(lib2)