+++ guppylang/CHANGELOG.md
@@ -14,6 +14,37 @@ This changelog documents user-facing changes to the Guppy language excluding cha
 
 As of August 2025, The Guppy language is undergoing rapid development and is currently unstable. There is a Guppy v1.0 stability [milestone](https://github.com/quantinuum/guppylang/milestone/12) that is a work in progress and subject to change.
 
+## [1.0.0-a5](https://github.com/Quantinuum/guppylang/compare/guppylang-v1.0.0-a4...guppylang-v1.0.0-a5) (2026-06-15)
+
+
+### Features
+
+* Add platform configuration to emulator  ([#1850](https://github.com/Quantinuum/guppylang/issues/1850)) ([d6bb388](https://github.com/Quantinuum/guppylang/commit/d6bb388ef590e8587042c1987334ca6fe34ce47f))
+* Add yy_max utility to sol standard library to match xx_max ([#1857](https://github.com/Quantinuum/guppylang/issues/1857)) ([d357302](https://github.com/Quantinuum/guppylang/commit/d3573029fbf5ebb1b4a060628f9f550933c32bdf))
+* Split qsystem into helios and sol modules ([#1848](https://github.com/Quantinuum/guppylang/issues/1848)) ([1c86e43](https://github.com/Quantinuum/guppylang/commit/1c86e43b71e703ec73f3560f0fdfec172704cc2e))
+
+
+### Bug Fixes
+
+* Use correct key for unitary flag metadata ([#1844](https://github.com/Quantinuum/guppylang/issues/1844)) ([4638bc1](https://github.com/Quantinuum/guppylang/commit/4638bc113b634dea1a04e30c08df162700b07bca))
+
+## [1.0.0-a4](https://github.com/Quantinuum/guppylang/compare/guppylang-v1.0.0-a3...guppylang-v1.0.0-a4) (2026-06-12)
+
+
+### ⚠ BREAKING CHANGES
+
+* Change the return type of measurement functions to a new `Measurement` type and remove opaque bools from the compiler ([#1707](https://github.com/Quantinuum/guppylang/issues/1707))
+* Make stdlib types state mutable with private fields ([#1840](https://github.com/Quantinuum/guppylang/issues/1840))
+
+### Features
+
+* Change the return type of measurement functions to a new `Measurement` type and remove opaque bools from the compiler ([#1707](https://github.com/Quantinuum/guppylang/issues/1707)) ([9460e6b](https://github.com/Quantinuum/guppylang/commit/9460e6b672fe84d87f360b7d521b6b9d5b736da2))
+
+
+### Code Refactoring
+
+* Make stdlib types state mutable with private fields ([#1840](https://github.com/Quantinuum/guppylang/issues/1840)) ([97c6962](https://github.com/Quantinuum/guppylang/commit/97c69624a1181804e8b2c1685941a394bae84d19))
+
 ## [1.0.0-a3](https://github.com/Quantinuum/guppylang/compare/guppylang-v1.0.0-a2...guppylang-v1.0.0-a3) (2026-06-10)